10_lesson/2al/requirements.txt
pytest==8.4.1
selenium==4.25.0
allure-pytest==2.15.0
allure-python-commons==2.15.0